requirements.txt
torch==1.10.2
torchvision==0.4.0
Pillow==8.3.2
argparse==1.4.0
tqdm==4.62.3
torchserve==0.5.2
numpy==1.19.5